# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FeedMill ERP API",
    "version": "2.0.0",
    "description": "Cattle Feed Manufacturing ERP System - REST API Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:3006/api",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/inventory/stock": {
      "get": {
        "summary": "Get stock summary",
        "tags": ["Inventory"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string" } },
          { "name": "godown_id", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Stock list" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}